requirements.txt
sounddevice
numpy
openai
anthropic
python-dotenv
